requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0
django-allauth==0.36.0
django-bootstrap4==0.0.6
django-heroku==0.3.1
djangorestframework==3.8.2
idna==2.7
markup==0.2
oauthlib==2.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-core==1.7.0
urllib3==1.23
whitenoise==3.3.1
